requirements.txt
ipykernel
python-dotenv
langchain_community
streamlit